Move header to proper location
